requirements.txt
alembic==1.3.0
amqp==2.5.2
beautifulsoup4==4.8.1
billiard==3.6.1.0
blinker==1.4
boto3==1.10.43
botocore==1.13.43
bs4==0.0.1
celery==4.3.0
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
Click==7.0
cryptography==2.8
decorator==4.0.11
docutils==0.15.2
fake-useragent==0.1.11
filetype==1.0.5
Flask==1.1.1
Flask-Assets==0.12
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-Moment==0.9.0
Flask-PageDown==0.2.2
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
idna==2.8
imageio==2.1.2
importlib-metadata==0.23
instagram-private-api==1.6.0.0
instagram-private-api-extensions==0.3.8
-e git+https://github.com/LevPasha/Instagram-API-python.git@cf0fe1d0167d153b0fc6c9be56e265e653c3da7a#egg=InstagramAPI
instapy-cli==0.0.12
itsdangerous==1.1.0
Jinja2==2.10.3
jmespath==0.9.4
kombu==4.6.6
lxml==4.4.2
Mako==1.1.0
MarkupSafe==1.1.1
more-itertools==7.2.0
moviepy==0.2.3.2
numpy==1.17.4
pandas==0.25.3
Pillow==6.2.1
pkg-resources==0.0.0
pycparser==2.19
PyJWT==1.7.1
python-dateutil==2.8.0
python-dotenv==0.10.3
python-editor==1.0.4
pytz==2019.3
redis==3.3.11
requests==2.11.1
requests-toolbelt==0.7.0
rq==1.1.0
s3transfer==0.2.1
six==1.13.0
soupsieve==1.9.5
SQLAlchemy==1.3.10
tqdm==4.11.2
urllib3==1.25.7
vine==1.3.0
webassets==0.12.1
Werkzeug==0.16.0
WTForms==2.2.1
zipp==0.6.0
asn1crypto==0.24.0
attrs==17.4.0
Automat==0.6.0
bcrypt==3.1.7
beautifulsoup4==4.8.1
blinker==1.4
bs4==0.0.1
certifi==2018.1.18
cffi==1.13.1
chardet==3.0.4
click==6.7
cloud-init==18.5
colorama==0.3.7
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
distro-info===0.18ubuntu0.18.04.1
httplib2==0.9.2
hyperlink==17.3.1
idna==2.6
incremental==16.10.1
Jinja2==2.10
jsonpatch==1.16
jsonpointer==1.10
jsonschema==2.6.0
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
MarkupSafe==1.0
netifaces==0.10.4
oauthlib==2.0.6
PAM==0.4.2
paramiko==2.6.0
plexcp==0.0.3
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.19
pycrypto==2.6.1
pygame==1.9.6
pygobject==3.26.1
PyJWT==1.5.3
PyNaCl==1.3.0
pyOpenSSL==17.5.0
pyserial==3.4
python-apt==1.6.4
python-debian==0.1.32
pyxdg==0.25
PyYAML==3.12
requests==2.18.4
requests-unixsocket==0.1.5
SecretStorage==2.3.1
service-identity==16.0.0
six==1.11.0
soupsieve==1.9.5
ssh-import-id==5.7
systemd-python==234
Twisted==17.9.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.22
virtualenv==16.7.9
zope.interface==4.3.2
